+++ edison-core/src/main/java/de/otto/edison/authentication/LdapRoleAuthenticationFilter.java
@@ -1,0 +1,62 @@
+package de.otto.edison.authentication;
+
+import de.otto.edison.authentication.configuration.LdapProperties;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Collection;
+
+import static java.util.Objects.requireNonNull;
+import static org.springframework.http.HttpHeaders.WWW_AUTHENTICATE;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
+
+/**
+ * Filter that checks for LDAP authentication with certain role once per request. Will not filter routes starting with
+ * {@link LdapProperties#whitelistedPaths}. Uses {@link LdapProperties} to set required role to access secured paths.
+ * Rejects requests with {@code HTTP 401} if authorization fails.
+ */
+public class LdapRoleAuthenticationFilter extends OncePerRequestFilter {
+
+    private static Logger LOG = LoggerFactory.getLogger(LdapRoleAuthenticationFilter.class);
+
+    private final Collection<String> whiteListedPaths;
+    private final String requiredRole;
+
+    public LdapRoleAuthenticationFilter(final LdapProperties ldapProperties) {
+        this.whiteListedPaths = requireNonNull(ldapProperties.getWhitelistedPaths(), "white listed paths must not be null");
+        this.requiredRole = requireNonNull(ldapProperties.getRequiredRole(), "required role must not be null");
+    }
+
+    @Override
+    protected boolean shouldNotFilter(final HttpServletRequest request) throws ServletException {
+        final String servletPath = request.getServletPath();
+        return whiteListedPaths
+                .stream()
+                .anyMatch(servletPath::startsWith);
+    }
+
+    @Override
+    protected void doFilterInternal(final HttpServletRequest request,
+                                    final HttpServletResponse response,
+                                    final FilterChain filterChain) throws ServletException, IOException {
+        if (request.isUserInRole(requiredRole)) {
+            LOG.debug("Found correct role for login."); // don't expose roles at successful login as this is a security issue
+            filterChain.doFilter(request, response);
+        } else {
+            LOG.warn("Did not found correct role for login.");
+            unauthorized(response);
+        }
+    }
+
+    void unauthorized(final HttpServletResponse httpResponse) {
+        httpResponse.addHeader(WWW_AUTHENTICATE, "Basic realm=Authorization Required");
+        httpResponse.setStatus(UNAUTHORIZED.value());
+    }
+
+}