+++ edison-core/src/test/java/de/otto/edison/authentication/configuration/LdapConfigurationTest.java
@@ -7,10 +7,13 @@ import org.junit.Test;
 import org.springframework.beans.factory.UnsatisfiedDependencyException;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.test.util.TestPropertyValues;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
+import java.util.ArrayList;
+
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.*;
 
 public class LdapConfigurationTest {
 
@@ -28,7 +31,7 @@ public class LdapConfigurationTest {
     }
 
     @Test
-    public void shouldRegisterLdapFilter() {
+    public void shouldRegisterLdapFilterButNotLdapRoleFilter() {
         this.context.register(EnableAutoConfig.class);
         TestPropertyValues
                 .of("edison.application.management.base-path=/internal")
@@ -40,6 +43,28 @@ public class LdapConfigurationTest {
         this.context.refresh();
 
         assertThat(this.context.containsBean("ldapAuthenticationFilter"), is(true));
+        assertThat(this.context.containsBean("ldapRoleAuthenticationFilter"), is(false));
+    }
+
+    @Test
+    public void shouldRegisterLdapFilterAndLdapRoleFilterWhenRoleIsRequired() {
+        this.context.register(EnableAutoConfig.class);
+        TestPropertyValues
+                .of("edison.application.management.base-path=/internal")
+                .and("edison.ldap.enabled=true")
+                .and("edison.ldap.host=localhost")
+                .and("edison.ldap.base-dn=test-dn")
+                .and("edison.ldap.required-role=someRole")
+                .and("edison.ldap.rdn-identifier=test-rdn")
+                .applyTo(context);
+        this.context.refresh();
+
+        assertThat(this.context.containsBean("ldapAuthenticationFilter"), is(true));
+        assertThat(this.context.containsBean("ldapRoleAuthenticationFilter"), is(true));
+
+        int ldapFilterOrder = this.context.getBean("ldapAuthenticationFilter", FilterRegistrationBean.class).getOrder();
+        int ldapRoleFilterOrder = this.context.getBean("ldapRoleAuthenticationFilter", FilterRegistrationBean.class).getOrder();
+        assertThat(ldapFilterOrder < ldapRoleFilterOrder, is(true));
     }
 
     @Test
@@ -110,5 +135,44 @@ public class LdapConfigurationTest {
         this.context.refresh();
 
         assertThat(this.context.getBean("ldapConnectionFactory").getClass().getSimpleName(), is(StartTlsLdapConnectionFactory.class.getSimpleName()));
+    }
+
+    @Test
+    public void shouldReadNewPrefixesPropertyAndIncludeOldPrefixProperty() {
+        this.context.register(EnableAutoConfig.class);
+        TestPropertyValues
+                .of("edison.application.management.base-path=/internal")
+                .and("edison.ldap.enabled=true")
+                .and("edison.ldap.host=localhost")
+                .and("edison.ldap.rdn-identifier=test-rdn")
+                .and("edison.ldap.base-dn=test-dn")
+                .and("edison.ldap.prefix=/deprecatedTestPrefix")
+                .and("edison.ldap.prefixes=/newTestPrefix")
+                .applyTo(context);
+        this.context.refresh();
+
+        FilterRegistrationBean filterRegistrationBean = this.context.getBean("ldapAuthenticationFilter", FilterRegistrationBean.class);
+        ArrayList<String> urlPatterns = new ArrayList<String>(filterRegistrationBean.getUrlPatterns());
+        assertThat(urlPatterns, hasSize(2));
+        assertThat(urlPatterns, containsInAnyOrder("/deprecatedTestPrefix/*", "/newTestPrefix/*"));
+    }
+
+    @Test
+    public void ensureBackwardsCompatibilityForPrefixesProperty() {
+        this.context.register(EnableAutoConfig.class);
+        TestPropertyValues
+                .of("edison.application.management.base-path=/internal")
+                .and("edison.ldap.enabled=true")
+                .and("edison.ldap.host=localhost")
+                .and("edison.ldap.rdn-identifier=test-rdn")
+                .and("edison.ldap.base-dn=test-dn")
+                .and("edison.ldap.prefix=/deprecatedTestPrefix")
+                .applyTo(context);
+        this.context.refresh();
+
+        FilterRegistrationBean filterRegistrationBean = this.context.getBean("ldapAuthenticationFilter", FilterRegistrationBean.class);
+        ArrayList<String> urlPatterns = new ArrayList<String>(filterRegistrationBean.getUrlPatterns());
+        assertThat(urlPatterns, hasSize(1));
+        assertThat(urlPatterns, containsInAnyOrder("/deprecatedTestPrefix/*"));
     }
 }